requirements.txt
joblib~=1.4.0
numpy~=2.2.2
matplotlib==3.9.2
gurobipy~=11.0.1
pyomo==6.8.0
z3-solver==4.13.3.0
pathos~=0.3.2
pandas==2.2.3
scikit-learn==1.5.2
seaborn~=0.13.2
intvalpy~=1.6.2
scipy~=1.15.1
pydoe3~=1.0.4
multiprocess~=0.70.16
psutil~=5.9.0
osqp~=0.6.7
pycddlib~=2.1.7
cvxpy~=1.6.0
setuptools
tqdm~=4.67.1